move: NotFoundException과 OwnershipException을 authorization 디렉토리에서 관리, 별도의 핸들러 생성
