--- eta/src/main/java/com/example/eta/exception/NotFoundException.java
@@ -1,11 +1,0 @@
-package com.example.eta.exception;
-
-public class NotFoundException extends RuntimeException {
-}
-
-// TODO: 별도의 클래스 파일로 분리할지, 아니면 Aspect 클래스 내부에 정의할지
-class PortfolioNotFoundException extends NotFoundException {
-}
-
-class RebalanacingNotFoundException extends NotFoundException {
-}
