+++ eta/src/main/java/com/example/eta/exception/authorization/NotFoundException.java
@@ -1,0 +1,4 @@
+package com.example.eta.exception.authorization;
+
+public class NotFoundException extends RuntimeException {
+}
